ball now launches in a random direction

+++ Source/Core/util.h
@@ -8,6 +8,6 @@ inline int random(int from, int to) {
 	return  rand() % (to+1) + from;
 }
 
-inline double random(double from, double to) {
-	return  from + static_cast <double> (rand()) / (static_cast <double> (RAND_MAX / (to - from)));
+inline float random(float from, float to) {
+	return  from + static_cast <float> (rand()) / (static_cast <float> (RAND_MAX / (to - from)));
 }
